% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{build_forecasts}
\alias{build_forecasts}
\title{Builds forecasts with the trained model.}
\usage{
build_forecasts(mod, n_pred, freq)
}
\arguments{
\item{mod}{The trained model.}

\item{n_pred}{The number of predictions}

\item{freq}{The frequency. Currently only daily data (1) and monthly data (12).
are supported.}
}
\value{
The predictions with nice date representations
}
\description{
Builds forecasts with the trained model.
}
