% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{calc_yardstick_metrics}
\alias{calc_yardstick_metrics}
\title{wrapper for calculating metrics with yardstick functions}
\usage{
calc_yardstick_metrics(forecasts, testset, func, detailed)
}
\arguments{
\item{forecasts}{The dataframe containing the forecasts.}

\item{testset}{The dataframe containing the testset.}

\item{func}{The yardstick function call.}

\item{detailed}{If True the metrics for each article will be returned.}
}
\value{
The calculated metrics for each method.
}
\description{
It is used in \code{\link{tf_calc_metrics}} for making the function calls for
different metrics modular.
}
\examples{
\dontrun{calc_yardstick_metrics(forecasts, testset, func = yardstick::mase, detailed)}
}
