% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{choose_dtl_metric}
\alias{choose_dtl_metric}
\title{The functions calcualtes one specific metric for all iterates.}
\usage{
choose_dtl_metric(forecasts, testset, metric = "rmse")
}
\arguments{
\item{forecasts}{The forecast dataframe}

\item{testset}{The testset dataframe}

\item{metric}{The metric which should be returned}
}
\value{
The calculated values for all iterates
}
\description{
The functions calcualtes one specific metric for all iterates.
}
\examples{
choose_dtl_metric(sales_forecast, sales_test, metric = "rmse")

}
