% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{create_forecasting_dates}
\alias{create_forecasting_dates}
\title{Creates a vector with the dates for which the forecasts should be created}
\usage{
create_forecasting_dates(data, n_pred)
}
\arguments{
\item{data}{The original dataset}

\item{n_pred}{number of the forecasts that should be produced}
}
\value{
The vector with the dates of the forecast horizon
}
\description{
Creates a vector with the dates for which the forecasts should be created
}
\examples{
\\donotrun{create_forecasting_dates(sales_data, 6)}
}
