% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypreprocess-package.R
\docType{package}
\name{efor-package}
\alias{efor}
\alias{efor-package}
\title{efor: efor: Easy Forecasts, a package for helping creating forecasts.}
\description{
The package efor provides a set of functions which help to work with multiple
timeseries. The package is providing a set of functions for
applying different preprocess and transformations methods to multiple
timeseries saved in a dataframe. Furthermore it helps modelling and evaluating
the resulting models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flostracke/tidypreprocess}
  \item Report bugs at \url{https://github.com/flostracke/tidypreprocess/issues}
}

}
\author{
\strong{Maintainer}: Florian Stracke \email{flostracke@gmail.com}

}
\keyword{internal}
