% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{forecast_and_sweep}
\alias{forecast_and_sweep}
\title{Actually calls the forecast functions and applies sweep to the result.}
\usage{
forecast_and_sweep(mod, n_pred)
}
\arguments{
\item{mod}{The pretrained model.}

\item{n_pred}{The number of predictions}
}
\value{
The produced forecast. The date column has to be corrected.
}
\description{
Actually calls the forecast functions and applies sweep to the result.
}
