% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{forecasts_timeseries}
\alias{forecasts_timeseries}
\title{Wrapper function for forecasts from the forecast package}
\usage{
forecasts_timeseries(data, func, n_pred, name, ...)
}
\arguments{
\item{data}{The dataset with all the articles}

\item{func}{The used forecast function}

\item{n_pred}{Number of the produced forecasts}

\item{...}{FUrther arguments which should be passed to the forecast function}
}
\value{
The produced forecasts
}
\description{
Wrapper function for forecasts from the forecast package
}
\examples{
\dontrun{forecasts_timeseries(sales_data, auto.arima, 6, "auto.arima")}
}
