% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utils.R
\name{get_subset}
\alias{get_subset}
\title{Gets a subet of n iterates of the dataset}
\usage{
get_subset(data, n)
}
\arguments{
\item{data}{The analysed dataframe}

\item{n}{Number of unique articles}
}
\value{
The dataframe with only n unique iterates
}
\description{
Gets a subet of n iterates of the dataset
}
\examples{
get_subset(sales_daily, n = 1)

}
