% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utils.R
\name{get_unique_iterates}
\alias{get_unique_iterates}
\title{Function returns a vector with the unique iterates in a dataset}
\usage{
get_unique_iterates(data)
}
\arguments{
\item{data}{The analysed dataset}
}
\value{
The vector with the unique iterates
}
\description{
Function returns a vector with the unique iterates in a dataset
}
\examples{
get_unique_iterates(sales_daily)

}
