% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_func_timefeatures.R
\name{tf_add_time_features}
\alias{tf_add_time_features}
\title{The function creates additional features from the date column. You can get
a stripped down version for monthly data with type = "monthly}
\usage{
tf_add_time_features(data, type = "all")
}
\arguments{
\item{data}{The used inpud data}

\item{type}{Can be "all" or "monthly". With type = "montly" you get a reduced
number of features for monthly data}
}
\value{
The dataframe with the additional features
}
\description{
The function creates additional features from the date column. You can get
a stripped down version for monthly data with type = "monthly
}
\examples{
sales_monthly \%>\%
            tf_add_time_features(type = "monthly")
}
