% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helpers.R
\name{tf_build_forecast}
\alias{tf_build_forecast}
\title{Function creates a dataframe with the forecasts. The forecasts will be
combined with the testset.}
\usage{
tf_build_forecast(v_forecasts, df_test, name)
}
\arguments{
\item{v_forecasts}{The created forecasts. You can pass a numeric vector or
a mlr Prediction object.}

\item{df_test}{The Testset which should be enriched with the forecasts.}

\item{name}{The name of the forecasting method.}
}
\value{
The Testset Dataframe, combined with the created forecasts.
}
\description{
Function creates a dataframe with the forecasts. The forecasts will be
combined with the testset.
}
