% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{tf_calc_metrics}
\alias{tf_calc_metrics}
\title{Calculates different metrics for multiple models.}
\usage{
tf_calc_metrics(forecasts, test, metrics = c("rmse", "mae", "rsq",
  "mase", "mape"), detailed = F)
}
\arguments{
\item{forecasts}{The dataframe containing the forecasts.}

\item{test}{The dataframe containing the testset.}

\item{metrics}{A list with the returned metrics. Following metrics are
               possible:

               - rmse
               - mae
               - rsq
               - mase
               - mape}

\item{detailed}{If True the rmse for each article will be returned.}
}
\value{
The calculated rmse for each method.
}
\description{
This function calculates different metrics for all contained forecasts in a dataframe.
The different models have to be seperated with the help of the key column.
}
\examples{
tf_calc_metrics(sales_forecast, sales_test)

}
