% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_func_tsclean.R
\name{tf_clean_grouped_ts}
\alias{tf_clean_grouped_ts}
\title{TSCLEAN for multiple articles}
\usage{
tf_clean_grouped_ts(data, parallel = T, freq = 12, ...)
}
\arguments{
\item{data}{The dataframe with the timeseries}

\item{parallel}{If False the function is not executed in parallel}

\item{...}{Further arguments that are passed to tsclen}
}
\value{
Returns the cleaned dataframe
}
\description{
Applie the tsclean function to multiple articles in a dataframe.
}
