% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_func_tsclean.R
\name{tf_clean_ts}
\alias{tf_clean_ts}
\title{Function for applying the forecast::tsclean function}
\usage{
tf_clean_ts(data, freq, ...)
}
\arguments{
\item{data}{The dataframe containing a single timeseries}

\item{freq}{The used frequence. 1 for daily data and 12 for monthly data.}

\item{...}{Further parameters passed to the forecast::tsclean function}
}
\value{
The cleaned timeseries in a tibble
}
\description{
Uses supsmu for non-seasonal series and a robust STL decomposition for seasonal #' #' series. To estimate missing values and outlier replacements, linear interpolation #' is used on the (possibly seasonally adjusted) series
}
