% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{tf_count_best_method}
\alias{tf_count_best_method}
\title{Counts how often each method is the best method against a testset.}
\usage{
tf_count_best_method(forecasts, testset, metric = "rmse")
}
\arguments{
\item{forecasts}{The dataframe with all the forecasts included}

\item{testset}{The testset for counting the best method}

\item{metric}{The specified metric for choosing the best method}
}
\value{
The dataframe with the information how many times each method has been the best method
}
\description{
Counts how often each method is the best method against a testset.
}
