% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_lag.R
\name{tf_create_lags}
\alias{tf_create_lags}
\title{Function can create}
\usage{
tf_create_lags(data, lag_var = y, n_lags = 1L:12L)
}
\arguments{
\item{data}{The dataframe which should be extended with lagged variables}

\item{lag_var}{The column name which should be lagged. You can add a vector
for lagging multiple variables at once}

\item{n_lags}{The number of lags that should be created. You can pass a
squence of natural numbers for creating multiple lags at once}
}
\value{
The dataframe with the added lagged variables
}
\description{
Function can create
}
\examples{
tf_create_lags(
            salesdata::sales_monthly,
            lag_var = y,
            n_lags = 1:6
          )
}
