% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{tf_create_model}
\alias{tf_create_model}
\title{Creates models based on ts objects.}
\usage{
tf_create_model(data, func, freq, ...)
}
\arguments{
\item{data}{The dataframe containing the timeseries.}

\item{func}{The used forecast method.}

\item{freq}{The used frequence. 1 for daily data and 12 for monthly data.}

\item{...}{More arguments specific to the used forecasting method.}
}
\value{
The fitted model for the specified method.
}
\description{
Helper function that creates models based on ts objects. For that the passed
dataframe is casted to a ts objeckt with the help of the timetk package.
You can use following methods with this function:
}
\details{
- all methods form the forecasts package
- forecastHybrid
- forecastxgb
- smooth package
}
