% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_decompose.R
\name{tf_decompose}
\alias{tf_decompose}
\title{Function calls anomalize::time_decompose() internal and decomposes all
series in the tibble}
\usage{
tf_decompose(data, ...)
}
\arguments{
\item{data}{A tibble object}

\item{...}{Possible parameters of the anomalize::time_decompose() function
Please refer to this function for more details.}
}
\value{
The decomposed timeseries in the tibble object
}
\description{
Function calls anomalize::time_decompose() internal and decomposes all
series in the tibble
}
\examples{
tf_decompose(sales_monthly)
}
