% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{tf_forecast}
\alias{tf_forecast}
\title{Creates forecast for a dateframe.}
\usage{
tf_forecast(data, n_pred, func, ...)
}
\arguments{
\item{data}{The dataframe containing the timeseries.}

\item{n_pred}{The forecast horizon.}

\item{func}{The used forecast method.}

\item{...}{More arguments specific to the used forecasting method.}
}
\value{
The forecasted values for the series
}
\description{
The Dataframe has to have the columny date and y. Following methods are
supported:
}
\details{
- all methods form the forecasts package
- prophet
- forecastHybrid
- forecastxgb
- smooth package
}
