% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{tf_get_best_method}
\alias{tf_get_best_method}
\title{Chooses the best method against a validation set from multiple forecasts}
\usage{
tf_get_best_method(forecasts_valid, valid_set, metric = "rmse")
}
\arguments{
\item{forecasts_valid}{The forecasts against the validation set}

\item{valid_set}{The validation set}

\item{metric}{The metric which should be used for obtaining the bestmodel}
}
\value{
A dataframe
}
\description{
Chooses the best method against a validation set from multiple forecasts
}
\examples{
tf_get_best_method(sales_forecast, sales_test, metric = "rmse")
}
