% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_func_boxcox.R
\name{tf_get_lambdas}
\alias{tf_get_lambdas}
\title{Get lambdas used in Boxcox Transformation.}
\usage{
tf_get_lambdas(data)
}
\arguments{
\item{data}{The data which the lambdas are obtained for. Needs a column
"iterate" as an unique id.}
}
\value{
Returns the lambdas for all the timeseries in the dataframe.
}
\description{
Gets the lambdas relevant for a boxcox transformation. The
lambdas are fitted on the data. The lambdas are returned for multiple
timeseries in a dataframe, if there is a column "iterate".
}
\examples{
tf_get_lambdas(sales_monthly)

}
