% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{tf_grouped_forecasts}
\alias{tf_grouped_forecasts}
\title{Creates forecast for a dateframe with mutlitple timeseries separted by the
iterate column.}
\usage{
tf_grouped_forecasts(data, n_pred, func, parallel = TRUE, ...)
}
\arguments{
\item{data}{The dataframe containing the timeseries.}

\item{n_pred}{The forecast horizon.}

\item{func}{The used forecast method.}

\item{parallel}{Specifies if the forecasts are created in parallel.}

\item{...}{More arguments specific to the used forecasting method.}
}
\value{
The forecasted values for the dataset.
}
\description{
The Dataframe has to have the columny date, iterate and y. Following methods
are supported:
}
\details{
\itemize{
    \item {forecasts package: All methods from the forecast packages
    (\url{https://CRAN.R-project.org/package=forecast}) should work}
    \item {prophet: (\url{https://CRAN.R-project.org/package=prophet})}
    \item {forecastHybrid: (\url{https://CRAN.R-project.org/package=forecastHybrid })}
    \item {smooth package: (\url{https://CRAN.R-project.org/package=smooth})}
}

The forecasts are created in parallel with the help of the furrr package
(\url{https://github.com/DavisVaughan/furrr}). Right now there is some bug
with the prophet model in parallel.Thats why you should create prophet
predictions with the parameter \code{parallel = FALSE}.
}
\examples{
\\donotrun{tf_grouped_forecasts(
           train_data,        # used training dataset
           n_pred = 6,        # number of predictions
           func = auto.arima, # used forecasting method
           parallel = TRUE    # for runiing in parallel
         )}
}
