% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_log.R
\name{tf_log}
\alias{tf_log}
\title{This functions applies a log transformation to multiple timeseries. There
has to be column y in the dataset.}
\usage{
tf_log(data, constant = 1e-05, ...)
}
\arguments{
\item{data}{The dataset on which the log transformation is applied to. The
values in the y column will be transformed}

\item{constant}{Before transforming there is a value added to the data. This
makes it possible to transform 0 values.}

\item{...}{Further arguments passed tp the log function.}
}
\value{
Return the log transformed data.
}
\description{
This functions applies a log transformation to multiple timeseries. There
has to be column y in the dataset.
}
\examples{
tf_log(sales_monthly)
}
