% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{tf_mean_forecast}
\alias{tf_mean_forecast}
\title{Returns a simple mean forecast for each iterate}
\usage{
tf_mean_forecast(data, h)
}
\arguments{
\item{data}{A tibble object}

\item{h}{The number of future dates that should be forecasted}
}
\value{
The tibble with the mean forecasts
}
\description{
Returns a simple mean forecast for each iterate
}
\examples{
tf_mean_forecast(sales_monthly, h = 2)
}
