% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plots.R
\name{tf_plot_n_forecasts}
\alias{tf_plot_n_forecasts}
\title{The functions plots the n best or worst forecasts for a given forecasting
method}
\usage{
tf_plot_n_forecasts(train_data, test_data, forecast_data,
  relevant_metric = "rmse", n = 3L, top_or_flop = "flop")
}
\arguments{
\item{top_or_flop}{}
}
\description{
The functions plots the n best or worst forecasts for a given forecasting
method
}
