% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plots.R
\name{tf_plot_preds_actuals}
\alias{tf_plot_preds_actuals}
\title{Plots the generated forecast against the actual values for each method. The
different forecast methods are grouped by a column "key".}
\usage{
tf_plot_preds_actuals(forecasts, test)
}
\arguments{
\item{forecasts}{The created forecasts for each method}

\item{test}{The testset with the actual values}
}
\description{
Plots the generated forecast against the actual values for each method. The
different forecast methods are grouped by a column "key".
}
\examples{
tf_plot_preds_actuals(sales_forecast, sales_test)
}
