% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{tf_plot_residuals}
\alias{tf_plot_residuals}
\title{Plots the distribtuion of the residuals of each method}
\usage{
tf_plot_residuals(testset, forecasts)
}
\arguments{
\item{testset}{The testset with the acutal values}

\item{forecasts}{The dataset with the forecasted values}
}
\description{
Plots the distribtuion of the residuals of each method
}
