% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prophet.R
\name{tf_prophet}
\alias{tf_prophet}
\title{Creates forecast with prophet for a dateframe.}
\usage{
tf_prophet(data, n_pred, freq, ...)
}
\arguments{
\item{data}{The dataframe containing the timeseries.}

\item{n_pred}{The forecast horizon.}

\item{freq}{The used frequency. 1 for daily data and 12 for monthly data.}

\item{...}{More arguments that are passed to the prophet model.}
}
\value{
The forecasted values for the series
}
\description{
The Dataframe has to have the columny date and y. This method is for creating
forecasts with the prophet model.
}
