% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{tf_prophet_grid}
\alias{tf_prophet_grid}
\title{Gridseach for Prophetmodels.}
\usage{
tf_prophet_grid(data, test_data, n_pred, freq, parallel = FALSE,
  parameter_grid, ...)
}
\arguments{
\item{data}{The Dataframe containing the data}

\item{n_pred}{The forecast horizon}

\item{parallel}{specifies if the forecasting is run in parallel}

\item{parameter_grid}{the used parameters in the random search}
}
\value{
A dataframe with the rmse for each parameter combination
}
\description{
Implements gridsearch for prophet models over multiple articles.
}
