% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_boxcox.R
\name{tf_remove_boxcox}
\alias{tf_remove_boxcox}
\title{Removes the boxcox transformation applied to multiple timeseries.}
\usage{
tf_remove_boxcox(data, lambdas)
}
\arguments{
\item{data}{The data with the boxcox transformation applied to.}

\item{lambdas}{The data frame with the lambdas. Can be created with
\code{\link{tf_get_lambdas}}}
}
\value{
Returns the series with the boxcox transformation removed.
}
\description{
Removes the boxcox transformation applied to multiple timeseries.
}
\examples{
lambdas <- tf_get_lambdas(sales_monthly)
          boxcox <- sales_monthly \%>\%
          tf_boxcox()
          tf_remove_boxcox(boxcox, lambdas)

}
