% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_log.R
\name{tf_remove_log}
\alias{tf_remove_log}
\title{This function rempves}
\usage{
tf_remove_log(data, constant = 1e-05)
}
\arguments{
\item{data}{The dataset on which the log transformation is removed. The
values in the y column will be transformed}

\item{constant}{Before applying the exp() function the constant will be
removed}
}
\value{
The dataset with the removed log transformation.
}
\description{
This function rempves
}
\examples{
salesdata::sales_monthly \%>\%
           tf_log() \%>\%
           tf_remove_log()
}
