% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_boxcox.R
\name{tf_boxcox}
\alias{tf_boxcox}
\title{Applies the boxcox transformation to multiple timeseries.}
\usage{
tf_boxcox(data)
}
\arguments{
\item{data}{The data which should be transformed.}
}
\value{
The data with boxcox transformed values.
}
\description{
Applies the boxcox transformation to multiple timeseries.
}
\examples{
tf_boxcox(salesdata::sales_monthly)

}
