% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_evaluation.R
\name{tf_calc_metrics}
\alias{tf_calc_metrics}
\title{Creates the metrics for mutliple models.}
\usage{
tf_calc_metrics(df_forecasts, df_test, metrics = list("rmse"),
  detailed = F)
}
\arguments{
\item{df_forecasts}{The dataframe containing the forecasts.}

\item{df_test}{The dataframe containing the testset.}

\item{metrics}{A list with the returned metrics. Following metrics are
               possible:

               - rmse
               - mae
               - rsq
               - mase
               - mape}

\item{detailed}{If True the rmse for each article will be returned.}
}
\value{
The calculated rmse for each method.
}
\description{
This function calculates the rmse for all contained forecasts in a dataframe.
The different models have to be seperated with the help of the key column.
}
\examples{
tf_calc_metrics(sales_forecast, sales_test)

}
