% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_forecasts.R
\name{tf_grouped_forecasts}
\alias{tf_grouped_forecasts}
\title{Creates forecast for a dateframe with mutlitple timeseries separted by the
iterate column.}
\usage{
tf_grouped_forecasts(data, n_pred, func, parallel = T, ...)
}
\arguments{
\item{data}{The dataframe containing the timeseries.}

\item{n_pred}{The forecast horizon.}

\item{func}{The used forecast method.}

\item{parallel}{Specifies if the forecasts are created in parallel.}

\item{...}{More arguments specific to the used forecasting method.}
}
\value{
The forecasted values for the dataset.
}
\description{
The Dataframe has to have the columny date, iterate and y. Following methods are
supported:
}
\details{
- all methods form the forecasts package
- prophet
- forecastHybrid
- forecastxgb
- smooth package

The forecasts are created in parallel with the help of the furrr package.
Since the package is quite fresh there might be some bugs involved
(for example the prophet model is somehow not working in parallel mode).
}
